Fixed the sequence of how the response is rendered. This is now a more logical flow.
When the request is accessing a Koowa component, and the method is GET and the format is HTML, Koowa will let Wordpress run its course which will go through the hooks and filters that manipulates the display. The component is dispatched but the response is not echoed. The response will be displayed when the admin-menu has been initiated.

When request is looking for a Koowa component and the method anything other than GET, Koowa will take over. If the format is not HTML (like json and rss), Koowa will also take over. Wordpress will be halted just after the "init" hook.