Get query information from table 'koowa_pages'.
This approach allows the component to be dispatched to the page it is attached to. This is run before headers are sent. The shortcode approach doesn't work because it is parsed after the headers are sent. So in the backend, there must be a custom way to attach koowa component views with a page. The query information is then saved at 'koowa_pages' table.