Divide the plugins into admin and site.
The new directory structure would be:

plugins/pluginname/admin/components/componentname
plugins/pluginname/site/components/componentname
plugins/pluginname/pluginname.php

Changes:
The namespace of plugins are the ones that are registered. If namespace is not registered look for it at plugins/koowa/components. Koowa and Application namespace are not registered.